+++ packages/angular/src/types.ts
@@ -48,3 +48,8 @@ export interface ValueAccessorConfig {
 export interface PackageJSON {
   types: string;
 }
+
+export interface ComponentInputProperty {
+  name: string;
+  required: boolean;
+}